Created basic rumble
Just did more cleanup of the rumble class and created a BasicRumble class instead
of the previously existing static method for createing the classic-style rumble patterns.
